successfully modularized firebase login functionality; renamed components;
