+++ .gitignore
@@ -23,4 +23,4 @@ yarn-debug.log*
 yarn-error.log*
 
 #firebase
-services/firestore.js
+*firestore.js
